Better handling of the access token expiration

+++ src/builders.rs
@@ -47,7 +47,7 @@ pub struct EsiBuilder {
     pub(crate) callback_url: Option<String>,
     pub(crate) scope: Option<String>,
     pub(crate) access_token: Option<String>,
-    pub(crate) access_expiration: Option<u64>,
+    pub(crate) access_expiration: Option<u128>,
     pub(crate) refresh_token: Option<String>,
     pub(crate) user_agent: Option<String>,
     pub(crate) http_timeout: Option<u64>,
@@ -98,7 +98,7 @@ impl EsiBuilder {
     }
 
     /// Set the access_expiration.
-    pub fn access_expiration(mut self, val: Option<u64>) -> Self {
+    pub fn access_expiration(mut self, val: Option<u128>) -> Self {
         self.access_expiration = val;
         self
     }
